🚚 api: [EVALUATION] move adapter to evaluation context
